requirements.txt
alembic
bcrypt
fastapi
psycopg2-binary
PyJWT
python-dotenv
SQLAlchemy
uvicorn
